Add seperate blueprint for user preferences
