+++ backend/requirements.txt
@@ -1,5 +1,6 @@
 beautifulsoup4==4.11.1
 bs4==0.0.1
+cachelib==0.9.0
 certifi==2022.9.24
 charset-normalizer==2.1.1
 click==8.1.3
@@ -7,6 +8,8 @@ colorama==0.4.6
 dnspython==2.2.1
 Flask==2.2.2
 Flask-Cors==3.0.10
+Flask-Login==0.6.2
+Flask-Session==0.4.0
 gunicorn==20.1.0
 html5lib==1.1
 idna==3.4
@@ -19,10 +22,13 @@ PyJWT==2.6.0
 pymongo==4.3.2
 python-dateutil==2.8.2
 python-dotenv==0.21.0
+python-http-client==3.3.7
 pytz==2022.6
 requests==2.28.1
+sendgrid==6.9.7
 six==1.16.0
 soupsieve==2.3.2.post1
+starkbank-ecdsa==2.2.0
 urllib3==1.26.12
 webencodings==0.5.1
 Werkzeug==2.2.2
